backend/requirements.txt
# Swing Trade Analyzer - Backend Dependencies
# Last Updated: Day 51 (February 11, 2026)

# Core Framework
flask>=2.0.0
flask-cors>=3.0.0

# Data Sources
yfinance>=0.2.0
defeatbeta-api>=0.1.0

# TradingView Screener (Day 11)
tradingview-screener==3.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0  # Day 44: Pattern detection (argrelextrema)

# Utilities
requests>=2.28.0
python-dotenv>=1.0.0          # Day 51: v4.14 - Load API keys from .env

# Multi-Source Data Providers (Day 51: v4.14)
pandas_datareader>=0.10.0     # Stooq fallback OHLCV provider

#scraping tools
stockanalysis-scraper==1.1.0
beautifulsoup4
selenium
webdriver-manager
